requirements.txt
arrow==0.15.4
backcall==0.1.0
decorator==4.4.1
IMAPClient==2.1.0
importlib-metadata==0.23
inflect==3.0.2
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
irc==17.1
jaraco.classes==2.0
jaraco.collections==2.1
jaraco.functools==2.0
jaraco.itertools==4.4.2
jaraco.logging==2.0
jaraco.stream==2.0
jaraco.text==3.1
jedi==0.15.1
more-itertools==7.2.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
schedule==0.6.0
six==1.13.0
tempora==1.14.1
termcolor==1.1.0
traitlets==4.3.3
wcwidth==0.1.7
zipp==0.6.0
